added operators
To keep tokens created to a minimum, any operators one after another will be grouped together. That means that '+' and '+=' are are both one item. This also means that operator chains like '+=!?<~>' are also one item and will be highlighted as such. This expects the code passed to it to be valid, compile-able code. I needed, I can make the changes to check for valid operators later